Got some basic dragging and dropping working

+++ four-fours-core/src/ui/has_location.rs
@@ -1,4 +1,4 @@
 pub trait HasLocation {
-  fn get_left(&self) -> i64;
-  fn get_top(&self) -> i64;
+  fn get_left(&self) -> f64;
+  fn get_top(&self) -> f64;
 }
