+++ four-fours-core/src/ui/has_mutable_size.rs
@@ -1,8 +1,8 @@
 
 pub trait HasMutableSize {
-  fn set_size_animated(&self, width: i64, height: i64, duraction_seconds: f64);
+  fn set_size_animated(&self, width: f64, height: f64, duraction_seconds: f64);
 
-  fn set_size(&self, width: i64, height: i64) {
+  fn set_size(&self, width: f64, height: f64) {
     self.set_size_animated(width, height, 0.0);
   }
 }
