+++ four-fours-core/src/ui/has_size.rs
@@ -1,4 +1,4 @@
 pub trait HasSize {
-  fn get_width(&self) -> i64;
-  fn get_height(&self) -> i64;
+  fn get_width(&self) -> f64;
+  fn get_height(&self) -> f64;
 }
